+++ src/PartnerCenter/Models/Subscriptions/Subscription.cs
@@ -18,9 +18,19 @@ namespace Microsoft.Store.PartnerCenter.Models.Subscriptions
     public sealed class Subscription : Contract
     {
         /// <summary>
+        /// Gets or sets a flag indicating whether or not action was taken.
+        /// </summary>
+        public bool ActionTaken { get; set; }
+
+        /// <summary>
         /// Gets or sets the actions.
         /// </summary>
         public IEnumerable<string> Actions { get; set; }
+
+        /// <summary>
+        /// Gets or sets a flag indicating whether or not attention is needed.
+        /// </summary>
+        public bool AttentionNeeded { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether automatic renew is enabled or not.
@@ -58,9 +68,9 @@ namespace Microsoft.Store.PartnerCenter.Models.Subscriptions
         public DateTime EffectiveStartDate { get; set; }
 
         /// <summary>
-        /// Gets or sets the entitlment identifier for the subscription.
+        /// Gets or sets the entitlement identifier for the subscription.
         /// </summary>
-        public string EntiltmentId { get; set; }
+        public string EntitlementId { get; set; }
 
         /// <summary>
         /// Gets or sets the subscription's friendly name.
@@ -113,7 +123,7 @@ namespace Microsoft.Store.PartnerCenter.Models.Subscriptions
         public string PartnerId { get; set; }
 
         /// <summary>
-        /// Gets or sets the publisher name.
+        /// Gets or sets the name of the publisher.
         /// </summary>
         public string PublisherName { get; set; }
 
@@ -138,8 +148,11 @@ namespace Microsoft.Store.PartnerCenter.Models.Subscriptions
         public IEnumerable<string> SuspensionReasons { get; set; }
 
         /// <summary>
-        /// Gets or sets the term duration.
+        /// Gets or sets the term duration. 
         /// </summary>
+        /// <remarks>
+        /// An ISO 8601 representation of the term's duration. The current supported values are P1M (1 month), P1Y (1 year) and P3Y (3 years).
+        /// </remarks>
         public string TermDuration { get; set; }
 
         /// <summary>