+++ src/PowerShell/Commands/ConnectPartnerCenter.cs
@@ -112,7 +112,7 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
 
             WriteDebug(string.Format(CultureInfo.CurrentCulture, Resources.ConnectPartnerCenterBeginProcess, ParameterSetName));
 
-            PartnerService.Instance.EnforceMfa = (EnforceMFA.IsPresent && EnforceMFA.ToBool()) ? true : false;
+            PartnerService.Instance.EnforceMfa = (EnforceMFA.IsPresent && EnforceMFA.ToBool());
 
             if (ParameterSetName.Equals(AccessTokenParameterSet, StringComparison.InvariantCultureIgnoreCase))
             {