+++ ChangeLog.md
@@ -75,6 +75,8 @@
     * TermDuration
 * Users
   * Corrected an issue with performing a query for users from a customer
+* Utilization
+  * Addressed an issue caused by the Partner Center API return a HTTP 204 no content when Azure utilization data is not yet ready in a dependent system.
 * Validations
   * Added the ability to request validation codes used to create Government Community Cloud customers